LiveKit-ai-translation/server/requirements.txt
livekit-agents
livekit-plugins-openai
python-dotenv
livekit-plugins-speechmatics
livekit-plugins-silero
openai
fastapi
uvicorn
websockets
aiohttp
supabase
asyncpg
httpx
pydantic